requirements.txt
pika==1.3.2
psycopg2-binary==2.9.3
pkg_resources==0.0.0
yfinance==0.2.18
APScheduler==3.10.1
Flask>=2.2.0
Werkzeug>=2.1.0